RM 6964: ARM: vmlinux.lds.S: Prepare infrastructure to support suspend to RAM feature the for STM32H7 targets:
The suspend/resume code needs to be located outside of SDRAM. It is
decided to use TCM on STM32H7-SOM. vmlinux.lds.S already contains the macro
to link code/data to the ITCM/DTCM, but it is placed under the HAVE_TCM
feature enable condition which implementation in Linux is specific to
the MMU targets.

It is not required to support all TCM features implemented in tcm.c for
STM32H7, instead the linker script will be updated by this commit to allow
to link code to the TCM areas even if HAVE_TCM is not enabled. For this,
new config options CONFIG_ITCM_BASE and CONFIG_DTCM_BASE must be used to
define the I/DTCM base addresses for the specific SoC, which can be
a no-MMU target in a common case.